+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - Inject ChannelManager inside CategoryReader
  - CategoryRepository::findOrderCategories takes a CategoryInterface
  - pim_base_connector.reader.doctrine.obsoleteproduct, pim_base_connector.reader.doctrine.obsoleteproduct.class, pim_magento_connector.reader.doctrine.product has been removed from readers.yml because they are unused
+ - Remove defaultLocale argument from OptionNormalizer::getOptionLabel
 
 # 1.1.20 (2015-02-20)
 ## Bug fixes